+++ axle-visualize/src/main/scala/axle/visualize/ScaledArea2D.scala
@@ -7,9 +7,9 @@ import spire.implicits.eqOps
 // http://www.apl.jhu.edu/~hall/java/Java2D-Tutorial.html
 
 case class ScaledArea2D[X, Y](
-    width: Int,
-    height: Int,
-    pad: Int,
+    width: Double,
+    height: Double,
+    pad: Double,
     minX: X,
     maxX: X,
     minY: Y,
@@ -24,8 +24,8 @@ case class ScaledArea2D[X, Y](
   val drawableWidth = width - (2 * pad)
   val drawableHeight = height - (2 * pad)
 
-  def framePoint(sp: Point2D[X, Y]): Point2D[Int, Int] = Point2D(
-    pad + (drawableWidth * lengthX.portion(minX, sp.x, maxX)).toInt,
-    height - pad - (drawableHeight * lengthY.portion(minY, sp.y, maxY)).toInt)
+  def framePoint(sp: Point2D[X, Y]): Point2D[Double, Double] = Point2D(
+    pad + (drawableWidth * lengthX.portion(minX, sp.x, maxX)),
+    height - pad - (drawableHeight * lengthY.portion(minY, sp.y, maxY)))
 
 }
